Update JWT to JWT_EXTENED

+++ resources/item.py
@@ -1,5 +1,7 @@
 from flask_restful import Resource, reqparse
-from flask_jwt import jwt_required
+from flask_jwt_extended import (jwt_required, get_jwt_claims,
+                                jwt_optional, get_jwt_identity,
+                                fresh_jwt_required)
 from models.item import ItemModel
 
 
@@ -19,13 +21,14 @@ class Item(Resource):
         help='Every item needs a store ID'
     )
 
-    @jwt_required()
+    @jwt_required
     def get(self, name):
         item = ItemModel.find_by_name(name)
         if item:
             return item.json()
         return {"message": "Item not found"}, 404
 
+    @fresh_jwt_required
     def post(self, name):
         if ItemModel.find_by_name(name):
             return {
@@ -42,7 +45,11 @@ class Item(Resource):
             }, 500  # internal server error
         return item.json(), 201
 
+    @jwt_required
     def delete(self, name):
+        claims = get_jwt_claims()
+        if not claims["is_admin"]:
+            return {"message": 'Admin privilege required'}, 401
         item = ItemModel.find_by_name(name)
         if item:
             item.delete_from_db()
@@ -62,5 +69,13 @@ class Item(Resource):
 
 
 class Itemlist(Resource):
+    @jwt_optional
     def get(self):
-        return {'items': list(map(lambda x: x.json(), ItemModel.query.all()))}
+        user_id = get_jwt_identity()
+        items = [item.json() for item in ItemModel.find_all()]
+        if user_id:
+            return {'item': items}, 200
+        return {
+            'items': [item['name'] for item in items],
+            'message': 'More data available if you log in'
+            }, 200
